Create separate patch route for updating passwords
This ensures that in case of a validation error, the user reaches the
same form that they were editing, and is a cleaner solution than hacking
into the project#update route.

Also adds the validations that ensure that users can't set the private
checkbox to true without also specifying a password.