Handle error when users incorrectly enter published_at
When users input invalid parameters to the published_at date fields,
Rails picks these up as `published_at(1i)`, `published_at(2i)` and
`published_at(3i)`. These then get passed directly to `.update`.

If any of these parameters are out of range, this triggers a
multiparameter assignment error, which is a 500.

It isn't possible to guard for these in the model, because the field is
of type date, and we can't create a valid date object from invalid
parameters.

Because of this, we have to validate in the controller, which is a bit
messy.