Set all images to public by default
Currently, images are private on DigitalOcean Spaces. This means that
embedding them in posts requires generating a one-time URL with access
controls.

Instead, we can make the images public by default, and migrate all
existing images to public using a rake task.

Note that public isn't the same as enumerable/guessable, as File listing
is still set to Restricted in the bucket settings. An example of a
public image is this:

https://design-history.ams3.digitaloceanspaces.com/8x3eyop5d3xd0ea5uax11lwlzkrx

So anyone can see it as long as they have that URL. But guessing that
UUID at the end isn't possible, except if you view a post. This seems
like a good trade-off for now, but we might consider adding "fully
private image hosting" later if deemed necessary.