Replace comments section turbo frame when submitting [-w]
This allows users to maintain their scroll position when they attempt to
submit a comment that has errors, and updates the list of comments with
their comment if their submission is successful.

The downside is that the form simply gets cleared with no notice that
their submission is successful. This should be revisited later.