Allow users to exclude images from list at bottom of post
This adds a new checkbox to each image that defaults to checked,
allowing users to opt out of showing certain images at the bottom of a
post.

It adds a migration to set `show_at_bottom: "1"` to all existing blobs.
All future images will be initialised with this to "1" by default.

Closes #320.