Sanitize post body and related links
This content is displayed using `.html_safe` and is vulnerable to XSS
injection. We should sanitize it on the way into the database to make it
easier for downstream libraries to handle it.

We have to take special care to work around code blocks, to allow users
to use arbitrary tags inside them without escaping.